refactor: add custom basemodel into utils of root
